replace rsa pkcs#11 header files with rewrite.
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@3066 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs11/pkcs11-display.c
@@ -680,7 +680,7 @@ const char *lookup_enum(CK_ULONG type, CK_ULONG value)
 
 void show_error( FILE *f, char *str, CK_RV rc )
 {
-  fprintf(f, "%s returned:  %ld %s", str, rc, lookup_enum ( RV_T, rc ));
+  fprintf(f, "%s returned:  %ld %s", str, (unsigned long) rc, lookup_enum ( RV_T, rc ));
   fprintf(f, "\n");
 }
